Changed several properties from float to int in SIF code.

+++ stonesoup/predictor/kalman.py
@@ -575,13 +575,13 @@ class StochasticIntegrationPredictor(KalmanPredictor):
         doc="The control model to be used. Default `None` where the predictor "
         "will create a zero-effect linear :class:`~.ControlModel`.",
     )
-    Nmax: float = Property(default=10, doc="maximal number of iterations of SIR")
-    Nmin: float = Property(
+    Nmax: int = Property(default=10, doc="maximal number of iterations of SIR")
+    Nmin: int = Property(
         default=5,
         doc="minimal number of iterations of stochastic integration rule (SIR)",
     )
     Eps: float = Property(default=5e-3, doc="allowed threshold for integration error")
-    SIorder: float = Property(
+    SIorder: int = Property(
         default=5, doc="order of SIR (orders 1, 3, 5 are currently supported)"
     )
